tests 001 and 002
