+++ bin/gaga_gauss_cond_test
@@ -1,10 +1,8 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-from shutil import copyfile
 import click
 from matplotlib import pyplot as plt
-from scipy.stats import gaussian_kde
 from gatetools import phsp
 import numpy as np
 import gaga_phsp as gaga
