+++ bin/gaga_gauss_plot
@@ -56,7 +56,7 @@ def gaga_gauss_plot(phsp_filename, pth_filename, n, m, epoch, x, y):
         print(cond.shape)
         fake = gaga.generate_samples3(params, G, m, cond=cond)
     else:
-        fake = gaga.generate_samples_no_cond(params, G, m, m, False, True)
+        fake = gaga.generate_samples_non_cond(params, G, m, m, False, True)
 
     # get 2D points
     x_ref = real[:, 0]
