+++ bin/gaga_generate
@@ -56,7 +56,7 @@ def gaga_generate(
     if cond_phsp is not None:
         fake = gaga.generate_samples3(params, G, n, cond=cond_data)
     else:
-        fake = gaga.generate_samples_no_cond(params, G, n, b, False, True)
+        fake = gaga.generate_samples_non_cond(params, G, n, b, False, True)
     end = time.time()
     elapsed = end - start
     pps = n / elapsed
