+++ gaga_phsp/__init__.py
@@ -3,6 +3,7 @@ from .gaga_helpers import *
 from .gaga_helpers_plot import *
 from .gaga_helpers_pet import *
 from .gaga_helpers_spect import *
+from .gaga_helpers_tests import *
 from .gaga_functions import *
 from .gaga import *
 from .gaga_model import *
